admin panel cleaned up
